# adminservicecontact/src/main/java/com/strongculture/service/contact/entity/system/GrantUserRoleReqVo.java
package com.strongculture.service.contact.entity.system;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

import java.util.List;

@Data
public class GrantUserRoleReqVo {
    @Schema(name = "用户id")
    private Long userId;
    @Schema(name = "角色id")
    private List<Long> roleIds;
}