# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台运营管理 API 文档",
    "description": "后台运营管理 Restful API",
    "version": "V1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8008/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/system/user/updatePassword": {
      "post": {
        "tags": [
          "管理员服务",
          "修改密码"
        ],
        "operationId": "updatePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/modify": {
      "post": {
        "tags": [
          "编辑管理员",
          "管理员服务"
        ],
        "operationId": "modifyUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/login": {
      "post": {
        "tags": [
          "用户登录",
          "管理员服务"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/user/list": {
      "post": {
        "tags": [
          "管理员服务",
          "查询管理员列表"
        ],
        "operationId": "listUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserListReqVo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListUserPO"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/user/grantUserRole": {
      "post": {
        "tags": [
          "给用户分配角色",
          "管理员服务"
        ],
        "operationId": "grantUserRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantUserRoleReqVo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/created": {
      "post": {
        "tags": [
          "管理员服务",
          "添加管理员"
        ],
        "operationId": "createdUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/user/remove": {
      "get": {
        "tags": [
          "管理员服务",
          "删除管理员"
        ],
        "operationId": "removeUser",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/login/validateCode": {
      "get": {
        "tags": [
          "发送图片验证码",
          "管理员服务"
        ],
        "operationId": "getCodeImage",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/getUserRoleByUserId": {
      "get": {
        "tags": [
          "获取管理员角色",
          "管理员服务"
        ],
        "operationId": "getUserRoleByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListRolePO"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/getById": {
      "get": {
        "tags": [
          "通过管理员id查询管理员信息",
          "管理员服务"
        ],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseUserPO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdatePasswordDTO": {
        "required": [
          "新密码",
          "旧密码",
          "用户账户"
        ],
        "type": "object",
        "properties": {
          "用户账户": {
            "type": "string"
          },
          "旧密码": {
            "type": "string"
          },
          "新密码": {
            "type": "string"
          }
        }
      },
      "BaseResponse": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "object"
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UserPO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "loginAccount": {
            "type": "string"
          },
          "loginPassword": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contactTel": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "okDel": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "role": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "memus": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        }
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "登录账户": {
            "type": "string"
          },
          "登录密码": {
            "type": "string"
          },
          "验证码": {
            "type": "string"
          }
        }
      },
      "UserListReqVo": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "loginAccount": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BaseResponseListUserPO": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserPO"
            }
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BaseResponseUserPO": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "$ref": "#/components/schemas/UserPO"
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "GrantUserRoleReqVo": {
        "type": "object",
        "properties": {
          "用户id": {
            "type": "integer",
            "format": "int64"
          },
          "角色id": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "RolePO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "okDel": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roleName": {
            "type": "string"
          },
          "roleRemark": {
            "type": "string"
          },
          "hasRole": {
            "type": "boolean"
          },
          "permission": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "BaseResponseListRolePO": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolePO"
            }
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "BaseResponseString": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "string"
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}